Time: 66 ms (48.85%) | Memory: 89.3 MB (12.21%) - LeetSync

+++ 3265-maximum-good-subarray-sum/maximum-good-subarray-sum.java
@@ -5,11 +5,11 @@ class Solution {
         for(int i = 0; i < n; i++) {
             prefix[i+1] = prefix[i] + nums[i];
         }
-        HashMap<Long, Integer> map = new HashMap<>();
+        HashMap<Integer, Integer> map = new HashMap<>();
         long ans = Long.MIN_VALUE;
         for(int i = 0; i < n; i++) {
-            long a = nums[i] - k;
-            long b = nums[i] + k;
+            int a = nums[i] - k;
+            int b = nums[i] + k;
             if(map.containsKey(a)) {
                 int idx = map.get(a);
                 ans = Math.max(ans, prefix[i+1] - prefix[idx]);
@@ -18,14 +18,14 @@ class Solution {
                 int idx = map.get(b);
                 ans = Math.max(ans, prefix[i+1] - prefix[idx]);
             }
-            if(map.containsKey((long)nums[i])) {
-                int idx = map.get((long)nums[i]);
+            if(map.containsKey(nums[i])) {
+                int idx = map.get(nums[i]);
                 if(prefix[idx] > prefix[i]) {
-                    map.put((long)nums[i], i);
+                    map.put(nums[i], i);
                 }
             }
             else {
-                map.put((long)nums[i], i);
+                map.put(nums[i], i);
             }
         }
         return ans == Long.MIN_VALUE?0:ans;
